chore: move payload builder types to standalone mod
