+++ crates/ethereum/node/src/lib.rs
@@ -23,3 +23,5 @@ pub use evm::{
 
 pub mod node;
 pub use node::EthereumNode;
+
+pub mod payload;
